Refactoring

+++ API/Controllers/ProductsController.cs
@@ -1,25 +1,57 @@
+using API.Dtos;
 using API.Errors;
 using API.Helpers;
+using AutoMapper;
+using Core.Entities;
 using Core.Interfaces;
-using Core.Models.Dto;
 using Core.Models.SearchObjects;
 using Core.Models.UpsertObjects;
-using Microsoft.AspNetCore.Authorization;
+using Infrastructure.Data;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace API.Controllers
 {
 
-    public class ProductsController : BaseCRUDController<ProductDto, ProductSearchObject, ProductUpsertObject, ProductUpsertObject>
+    public class ProductsController : BaseCRUDController
+        <ProductDto, Product, ProductSearchObject, ProductUpsertObject, ProductUpsertObject>
     {
         private readonly IProductService _service;
         private readonly IPhotoService _photoService;
+        private readonly BarbershopContext _context;
+        private readonly IMapper _mapper;
 
-        public ProductsController(IProductService service, IPhotoService photoService)
-            : base(service)
+        public ProductsController(BarbershopContext context, IProductService service, IPhotoService photoService, IMapper mapper)
+            : base(service, mapper)
         {
+            _mapper = mapper;
+            _context = context;
             _service = service;
             _photoService = photoService;
+        }
+
+        // out for development
+        // [Authorize(Roles = "Admin")] 
+        [HttpDelete("{id}")]
+        public override async Task<ActionResult> Delete(int id)
+        {
+            var product = await _service.GetByIdAsync(id);
+
+            foreach (var photo in product.Photos)
+            {
+                if (photo.Id > 16)
+                {
+                    _photoService.DeleteFromDisk(photo);
+                    _context.Photos.Remove(photo);
+                }
+            }
+
+            if (!await _service.Delete(id))
+            {
+                return BadRequest(new ApiResponse(400, "Problem deleting product"));
+            }
+
+            return Ok();
         }
 
         // out for development
@@ -27,17 +59,30 @@ namespace API.Controllers
         [HttpPut("{id}/photo")]
         public async Task<ActionResult<ProductDto>> AddProductPhoto(int id, [FromForm] PhotoInputModel uploadPhotoFile)
         {
+            var product = await _service.GetByIdAsync(id);
+
             if (uploadPhotoFile.Photo.Length > 0)
             {
                 var photo = await _photoService.SaveToDiskAsync(uploadPhotoFile.Photo);
 
                 if (photo != null)
                 {
-                    var productDto = await _service.AddProductPhoto(id, photo);
-                    return productDto;
+                    product.AddPhoto(photo.PictureUrl, photo.FileName);
+
+                    _context.Products.Attach(product);
+                    _context.Entry(product).State = EntityState.Modified;
+
+                    var result = await _context.SaveChangesAsync();
+
+                    if (result <= 0) return BadRequest(new ApiResponse(400, "Problem adding photo product"));
+                }
+                else
+                {
+                    return BadRequest(new ApiResponse(400, "problem saving photo to disk"));
                 }
             }
-            return BadRequest(new ApiResponse(400, "Problem adding product photo"));
+
+            return _mapper.Map<Product, ProductDto>(product);
         }
 
         // out for development
@@ -45,22 +90,34 @@ namespace API.Controllers
         [HttpDelete("{id}/photo/{photoId}")]
         public async Task<ActionResult> DeleteProductPhoto(int id, int photoId)
         {
-            try
-            {
-                var photo = await _service.DeleteProductPhoto(id, photoId);
+            var product = await _service.GetByIdAsync(id);
 
-                if (photo != null)
-                {
-                    _photoService.DeleteFromDisk(photo);
-                    return Ok();
-                }
+            var photo = product.Photos.SingleOrDefault(x => x.Id == photoId);
 
-                return BadRequest(new ApiResponse(400, "Problem deleting product photo"));
-            }
-            catch (Exception ex)
+            if (photo != null)
             {
-                return BadRequest(new ApiResponse(400, ex.Message));
+                if (photo.IsMain)
+                    return BadRequest(new ApiResponse(400,
+                        "You cannot delete the main photo"));
+
+                _photoService.DeleteFromDisk(photo);
+                _context.Remove(photo);
             }
+            else
+            {
+                return BadRequest(new ApiResponse(400, "Photo does not exist"));
+            }
+
+            product.RemovePhoto(photoId);
+
+            _context.Products.Attach(product);
+            _context.Entry(product).State = EntityState.Modified;
+
+            var result = await _context.SaveChangesAsync();
+
+            if (result <= 0) return BadRequest(new ApiResponse(400, "Problem adding photo product"));
+
+            return Ok();
         }
 
         // out for development
@@ -68,21 +125,20 @@ namespace API.Controllers
         [HttpPatch("{id}/photo/{photoId}")]
         public async Task<ActionResult<ProductDto>> SetMainPhoto(int id, int photoId)
         {
-            try
-            {
-                var productDto = await _service.SetProductMainPhoto(id, photoId);
+            var product = await _service.GetByIdAsync(id);
 
-                if (productDto != null)
-                {
-                    return Ok(productDto);
-                }
+            if (product.Photos.All(x => x.Id != photoId)) return NotFound(new ApiResponse(404));
 
-                return BadRequest(new ApiResponse(400, "Problem setting product main photo"));
-            }
-            catch (Exception ex)
-            {
-                return NotFound(new ApiResponse(404, ex.Message));
-            }
+            product.SetMainPhoto(photoId);
+
+            _context.Products.Attach(product);
+            _context.Entry(product).State = EntityState.Modified;
+
+            var result = await _context.SaveChangesAsync();
+
+            if (result <= 0) return BadRequest(new ApiResponse(400, "Problem adding photo product"));
+
+            return _mapper.Map<Product, ProductDto>(product);
         }
     }
 }
